+++ assets/js/live_motion/types.ts
@@ -1,4 +1,5 @@
 import { AnimationOptionsWithOverrides, MotionKeyframesDefinition, spring } from '@motionone/dom';
+import { MotionState, VariantDefinition } from 'motion';
 
 export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
 
@@ -11,6 +12,7 @@ type LiveViewJSDefinition = any[];
 // type SpringOptions is not exported, so we extract it from the function definition.
 export type LiveMotionSpringOptions = Parameters<typeof spring>[0];
 
+// TODO: check if can be removed or if we can get away someway else
 export type LiveMotionAnimationOptions = AnimationOptionsWithOverrides & {
   __easing?: ['spring', LiveMotionSpringOptions];
 };
@@ -22,9 +24,10 @@ export type LiveMotionConfigOptions = {
 };
 
 export type LiveMotionConfig = {
-  keyframes: MotionKeyframesDefinition;
+  initial: VariantDefinition;
+  animate: VariantDefinition;
   transition?: LiveMotionAnimationOptions;
-  exit?: MotionKeyframesDefinition;
+  exit?: VariantDefinition;
   opts: LiveMotionConfigOptions;
 };
 
@@ -32,46 +35,49 @@ export type LiveMotionHooksDefinition = {
   Motion: LiveMotionHook;
 };
 
-export type LiveMotionHook = ThisType<{
+// TODO: Check if we can use Options form Motion directly
+export type MotionOptions = Pick<LiveMotionConfig, 'initial' | 'animate' | 'exit' | 'transition'>;
+
+export type MaybeAnimateOptions = {
+  force: boolean;
+};
+
+type LiveViewHookType = {
   // LiveView provided functions
+  mounted?(): void;
+  updated?(): void;
+  destroyed?(): void;
+  disconnected?(): void;
+  reconnected?(): void;
+
   el: HTMLElement;
   liveSocket: any;
-  mounted(): any;
-  updated(): any;
-  destroyed(): any;
-  disconnected(): any;
-  reconnected(): any;
+};
 
+type LiveMotionType = {
+  eventHandlers: Record<string, () => void>;
   getConfig(): LiveMotionConfig | undefined;
-  maybeAnimate(): void;
-}>;
+  getMotionOptions(): MotionOptions | undefined;
+  maybeAnimate(options?: MaybeAnimateOptions): void;
+  state?: MotionState;
+  cleanup?: ReturnType<MotionState['mount']>;
+};
 
-export interface LiveMotionEvent<T> extends Event {
+export type LiveMotionHook = LiveMotionType &
+  Pick<LiveViewHookType, 'el' | 'mounted' | 'destroyed' | 'updated'> &
+  ThisType<LiveMotionType & LiveViewHookType>;
+
+export interface LiveMotionEvent<T = unknown> extends Event {
   readonly detail?: T;
   target: HTMLElement;
 }
 
-export type LiveMotionAnimateEvent = LiveMotionEvent<
-  Pick<LiveMotionConfig, 'keyframes' | 'transition'>
->;
+export type LiveMotionAnimateEvent = LiveMotionEvent;
 
-export type LiveMotionHideEvent = LiveMotionEvent<LiveMotionConfig>;
+export type LiveMotionHideEvent = LiveMotionEvent;
 
-export type LiveMotionShowEvent = LiveMotionEvent<
-  Pick<LiveMotionConfig, 'keyframes' | 'transition'> & {
-    display: string;
-  }
->;
+export type LiveMotionShowEvent = LiveMotionEvent<{
+  display: string;
+}>;
 
-export type LiveMotionToggleEvent = LiveMotionEvent<
-  Pick<LiveMotionConfig, 'keyframes' | 'transition'> & {
-    keyframes: {
-      in: LiveMotionConfig['keyframes'];
-      out: LiveMotionConfig['keyframes'];
-    };
-    transition: {
-      in: LiveMotionConfig['transition'];
-      out: LiveMotionConfig['transition'];
-    };
-  }
->;
+export type LiveMotionToggleEvent = LiveMotionEvent;